docker compose up starts service and db

+++ internal/nrtm4serve/rpc/rpchandler.go
@@ -47,7 +47,7 @@ type JSONRPCRequest struct {
 
 // JSONRPCError An error
 type JSONRPCError struct {
-	Code    int64  `json:"code"`
+	Code    int    `json:"code"`
 	Message string `json:"message"`
 }
